requirements.txt
Flask
PyPDF2
openai
python-dotenv
flask-cors
numpy
pandas
scikit-learn
requests